+++ src/pages/Detail/index.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import { View } from 'react-native';
+
+export default function Detail() {
+ return (
+   <View/>
+  );
+}
